fix: 课程下需至少保留一个教学计划

+++ src/Biz/Course/Service/Impl/CourseServiceImpl.php
@@ -348,6 +348,10 @@ class CourseServiceImpl extends BaseService implements CourseService
         if (!empty($subCourses)) {
             throw $this->createAccessDeniedException('该教学计划在班级下存在引用，请先删除相关引用');
         }
+        $courseCount = $this->getCourseDao()->count(array('courseSetId' => $course['courseSetId']));
+        if($courseCount <= 1){
+            throw $this->createAccessDeniedException('课程下至少需保留一个教学计划');
+        }
         return $this->getCourseDeleteService()->deleteCourse($id);
     }
